+++ Lib/xTools4/dialogs/variable/GlyphValidator.py
@@ -56,7 +56,8 @@ def validationGroupFactory(glyph, defaultGlyph=None):
     else:
         # contours equal to default
         if checkResults['compatibility']['points'] and checkResults['equality']['points']:
-            if glyph.font.path != defaultGlyph.font.path:
+            # if glyph.font.path != defaultGlyph.font.path 
+            if glyph.width == defaultGlyph.width:
                 validationGroup = 'contoursEqual'
             else:
                 validationGroup = 'contoursDifferent'