+++ xTools4.roboFontExt/lib/menus/batch/build-glyphs.py
@@ -1,5 +1,9 @@
 # menuTitle : build glyphs
 
+from importlib import reload
+import xTools4.dialogs.batch.build
+reload(xTools4.dialogs.batch.build)
+
 from mojo.roboFont import OpenWindow
 from xTools4.dialogs.batch.build import BatchBuildGlyphsDialog